Permit larger anti-aliased fonts
Size limit for anti-aliased fonts raised to ~200 pixels

+++ Extensions/Smooth_font.h
@@ -35,7 +35,7 @@ fontMetrics gFont = { 0, 0, 0, 0, 0, 0, 0 };
   uint8_t*  gHeight = NULL;   //cheight
   uint8_t*  gWidth = NULL;    //cwidth
   uint8_t*  gxAdvance = NULL; //setWidth
-  int8_t*   gdY = NULL;       //topExtent
+  int16_t*  gdY = NULL;       //topExtent
   int8_t*   gdX = NULL;       //leftExtent
   uint32_t* gBitmap = NULL;   //file pointer to greyscale bitmap
